add some options of warning in CMakeLists

+++ src/util/ush_string.h
@@ -3,7 +3,7 @@
 
 #include "stddef.h"
 
-void ush_itoa(char *str, int num);
+void ush_itoa(char *str, long num);
 
 // conbine a string to a num
 void ush_string_certname(char *certname, size_t sz, const char *name, int cert);
